AMD-793 [frontend] refactor: abstract timeout alert component for config page
