+++ frontend/src/containers/ConfigStep/Board/index.tsx
@@ -6,12 +6,12 @@ import {
 } from '@src/components/Common/ConfigForms';
 import { updateShouldGetBoardConfig } from '@src/context/Metrics/metricsSlice';
 import { KEYS, useVerifyBoardEffect } from '@src/hooks/useVerifyBoardEffect';
+import { StyledAlert } from '@src/containers/ConfigStep/TimeoutAlert/style';
 import { ConfigButtonGrop } from '@src/containers/ConfigStep/ConfigButton';
 import { useAppSelector, useAppDispatch } from '@src/hooks/useAppDispatch';
 import { InputLabel, ListItemText, MenuItem, Select } from '@mui/material';
 import { ConfigSelectionTitle } from '@src/containers/MetricsStep/style';
 import { selectIsBoardVerified } from '@src/context/config/configSlice';
-import { StyledAlert } from '@src/containers/ConfigStep/Board/style';
 import { BOARD_TYPES, CONFIG_TITLE } from '@src/constants/resources';
 import HighlightOffIcon from '@mui/icons-material/HighlightOff';
 import { Loading } from '@src/components/Loading';
